Split view tests to match their tested modules.
